Document the expected Swift version support for packages
## Motivation

The `swiftlang` project contains packages that are consumed by the ecosystem via the Swift Package Manager. For a healthy ecosystem these packages need to support multiple Swift version to allow for a migration period when a new Swift version is released.

## Modifications

This PR expands on the expectations for packages that are part of `swiftlang`. In particular, it adds defines the expected Swift versions and platforms supported by packages. While doing so it clearly separates projects into projects that are released as part of the Swift Toolchain and projects that are released as Swift packages. The version and platform support only applies to the latter.

## Result

We now have a central place that defines the expectation around Swift and platform support for packages that we can reference across the `swiftlang` organizations. Packages outside of the `swiftlang` organization can choose to follow these expectations as well but are not obliged to do so.